+++ CHANGES.md
@@ -5,6 +5,7 @@
   - fixed transparency depth sorting when using an orthographic camera. the original math was so scuffed (thanks Ben)
   - fix crash caused by https://github.com/ccrma/chugl/issues/8 (thanks Nathan)
   - calling Material.storageBuffer(...) with an empty chuck array no longer crashes
+  - fix GWindow.sizeLimits(...) to allow non-integer aspect ratios, e.g. 1.6:1
 - New Examples
   - deep/meshlines.ck: 3D line rendering that supports per-vertex line width, dashed lines, and more